auth_server/requirements.txt
PyJWT==1.3.1
Flask==2.3.2
python-dotenv==1.0.0
